Hiep

+++ src/main/java/vn/diemdanh/hethong/entity/DiemDanhLog.java
@@ -18,7 +18,7 @@ public class DiemDanhLog {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
@@ -27,7 +27,7 @@ public class DiemDanhLog {
     private DiemDanh maDd;
 
     @Column(name = "lan_diem_danh", columnDefinition = "int UNSIGNED not null")
-    private Long lanDiemDanh;
+    private Integer lanDiemDanh;
 
     @ColumnDefault("CURRENT_TIMESTAMP")
     @Column(name = "thoi_gian_diem_danh")
